suggested fixes
